Model code is now working well.
Also showing root note on the instruments now.

Work to do:
- Implement more choices in the form to show more modes and chords.
- Do something with scales and chords that have double sharps/flats.
Either implement double sharp/flat functionality or show a message
(maybe pointing to the same scale/chord with a different name). Atm it
is going to error 500 page.
- Dynamically creating the options in the form and setting 'selected'
after initial choice. Now it always reverts back to C ionian, which is
confusing.